arch: arm64: change the formal parameters to fit the new kernel data structure
After commit 46ddfb9d1e46 (arm64: Treat ESR_ELx as a 64-bit register),
ESR_ELx is defined as 64-bit register. We need to adapt this change for
marvell erratum 38891 to avoid the following compilation problems.

arch/arm64/kernel/traps.c:702:28: error: initialization of 'void (*)(long unsigned int,  struct pt_regs *)' from incompatible pointer type 'void (*)(unsigned int,  struct pt_regs *)' [-Werror=incompatible-pointer-types]
  702 |                 .handler = user_cache_maint_handler,
      |                            ^~~~~~~~~~~~~~~~~~~~~~~~
arch/arm64/kernel/traps.c:702:28: note: (near initialization for 'sys64_hooks[0].handler')
arch/arm64/kernel/traps.c:708:28: error: initialization of 'void (*)(long unsigned int,  struct pt_regs *)' from incompatible pointer type 'void (*)(unsigned int,  struct pt_regs *)' [-Werror=incompatible-pointer-types]
  708 |                 .handler = ctr_read_handler,
      |                            ^~~~~~~~~~~~~~~~

+++ arch/arm64/kernel/traps.c
@@ -539,7 +539,7 @@ NOKPROBE_SYMBOL(do_ptrauth_fault);
 	}
 
 void
-user_cache_maint_handler_errtaum_38891(unsigned int esr, struct pt_regs *regs)
+user_cache_maint_handler_errtaum_38891(unsigned long esr, struct pt_regs *regs)
 {
 	unsigned long address;
 	int rt = ESR_ELx_SYS64_ISS_RT(esr);
@@ -590,7 +590,7 @@ user_cache_maint_handler_errtaum_38891(unsigned int esr, struct pt_regs *regs)
 }
 
 
-void user_cache_maint_handler_generic(unsigned int esr, struct pt_regs *regs)
+void user_cache_maint_handler_generic(unsigned long esr, struct pt_regs *regs)
 {
 	unsigned long tagged_address, address;
 	int rt = ESR_ELx_SYS64_ISS_RT(esr);
@@ -630,7 +630,7 @@ void user_cache_maint_handler_generic(unsigned int esr, struct pt_regs *regs)
 		arm64_skip_faulting_instruction(regs, AARCH64_INSN_SIZE);
 }
 
-static void user_cache_maint_handler(unsigned int esr, struct pt_regs *regs)
+static void user_cache_maint_handler(unsigned long esr, struct pt_regs *regs)
 {
 	if (cpus_have_cap(ARM64_WORKAROUND_MARVELL_38891))
 		user_cache_maint_handler_errtaum_38891(esr, regs);
@@ -638,7 +638,7 @@ static void user_cache_maint_handler(unsigned int esr, struct pt_regs *regs)
 		user_cache_maint_handler_generic(esr, regs);
 }
 
-static void ctr_read_handler(unsigned int esr, struct pt_regs *regs)
+static void ctr_read_handler(unsigned long esr, struct pt_regs *regs)
 {
 	int rt = ESR_ELx_SYS64_ISS_RT(esr);
 	unsigned long val = arm64_ftr_reg_user_value(&arm64_ftr_reg_ctrel0);
